requirements.txt
termcolor==2.3.0
openai==1.55.3
langchain==0.0.237
pydantic==1.10.9
typing-extensions>=4.11,<5
google-generativeai==0.2.0
colorlog==6.7.0
redis==5.0.0
